+++ front-end/src/components/pageLogin/FormLogin.jsx
@@ -1,0 +1,23 @@
+import React, { useContext } from 'react';
+import LoginContext from '../../context/LoginContext';
+import InputsForm from './InputsForm';
+import ButtonsForm from './ButtonsForm';
+
+function FormLogin() {
+  const {
+    dataUser: user,
+    hadleIputs: handleChange,
+    handleButton: handleClick,
+    isDisabled: valid,
+    router: history,
+  } = useContext(LoginContext);
+
+  return (
+    <div>
+      { InputsForm(user, handleChange) }
+      { ButtonsForm(valid, handleClick, history) }
+    </div>
+  );
+}
+
+export default FormLogin;
